+++ src/lib/statistics/chunk_statistics/histograms/equal_distinct_count_histogram.hpp
@@ -46,7 +46,7 @@ class EqualDistinctCountHistogram : public AbstractHistogram<T> {
   EqualDistinctCountHistogram(const std::vector<std::string>&& bin_minima, const std::vector<std::string>&& bin_maxima,
                               const std::vector<HistogramCountType>&& bin_heights,
                               const HistogramCountType distinct_count_per_bin, const BinID bin_count_with_extra_value,
-                              const std::string& supported_characters, const uint32_t string_prefix_length);
+                              const std::string& supported_characters, const size_t string_prefix_length);
 
   /**
    * Create a histogram based on the data in a given segment.
